Add CPU pinning support (#8)
Enable libvirt CPU pinning by setting the LibvirtCPUPinSet parameter
in environments/numa.yaml and adding it to the list of deploy
templates.

This will also change the compute image name, so that isolcpus
can be set for that image. This is done via the deploy script
as it requires setting a kernel parameter which is done via
modification of grub.conf in the image for compute nodes.