Fixes issue where ping fails after os-net-config (#31)
Issue where os-collect-config tries to execute a heat script that pings
each IP.  This fails because we never move the tenant nic to br-phy
(becuase we are going to bind it to dpdk).  The new behavior is we add
the port to br-phy first (so ping passes) and then we remove it before
we bind.