Default tenant_nic to its physical name (#88)
Currently we assume tenant_nic is a virtual interface name (like nic1),
so we perform an extra hiera lookup, this is not needed if the user
specifies physical nic name in network_settings file. This patch defaults
the hiera lookup of the nic interface lookup to be what the user specifies
if a hiera value can't be found.